moved GroupModelForm into new file 'forms/group.py'
