+++ base/forms/education_group/create.py
@@ -26,11 +26,9 @@
 from django import forms
 
 from base.forms.learning_unit.entity_form import EntitiesVersionChoiceField
-from base.models import campus, education_group_type, group_element_year, academic_year
+from base.models import campus, education_group_type, group_element_year
 from base.models.education_group import EducationGroup
-from base.models.education_group_year import EducationGroupYear
 from base.models.entity_version import find_main_entities_version, get_last_version
-from base.models.enums import education_group_categories
 
 
 class MainTeachingCampusChoiceField(forms.ModelChoiceField):
@@ -71,34 +69,6 @@ def _save_group_element_year(parent, child):
     # TODO :: what if this relation parent/child already exists? Should we create a new GroupElementYear anymore?
     if parent:
         group_element_year.get_or_create_group_element_year(parent, child)
-
-
-class GroupModelForm(forms.ModelForm):
-
-    class Meta:
-        model = EducationGroupYear
-        fields = ("acronym", "partial_acronym", "education_group_type", "title", "title_english", "credits",
-                  "main_teaching_campus", "academic_year", "remark", "remark_english", "min_credits", "max_credits",
-                  "administration_entity")
-        field_classes = {
-            "administration_entity": MainEntitiesVersionChoiceField,
-            "main_teaching_campus": MainTeachingCampusChoiceField
-        }
-
-    def __init__(self, *args, **kwargs):
-        self.parent = kwargs.pop("parent", None)
-        super().__init__(*args, **kwargs)
-        _init_education_group_type_field(self.fields["education_group_type"],
-                                         self.parent,
-                                         education_group_categories.GROUP)
-        _init_academic_year(self.fields["academic_year"], self.parent)
-
-        _preselect_entity_version_from_entity_value(self)
-
-    def save(self, *args, **kwargs):
-        education_group_year = super().save(*args, **kwargs)
-        _save_group_element_year(self.parent, education_group_year)
-        return education_group_year
 
 
 class EducationGroupModelForm(forms.ModelForm):
@@ -145,15 +115,3 @@ class CommonBaseForm:
         for form in self.forms.values():
             errors.update(form.errors)
         return errors
-
-
-class GroupForm(CommonBaseForm):
-
-    def __init__(self, data, instance=None, parent=None):
-        educ_group_year_form = GroupModelForm(data, instance=instance, parent=parent)
-
-        education_group = instance.education_group if instance else None
-
-        educ_group_model_form = EducationGroupModelForm({}, instance=education_group)
-
-        super(GroupForm, self).__init__(educ_group_year_form, educ_group_model_form)
