+++ base/tests/views/education_groups/test_update.py
@@ -30,7 +30,7 @@ from django.test import TestCase
 from django.urls import reverse
 from waffle.testutils import override_flag
 
-from base.forms.education_group.create import GroupModelForm
+from base.forms.education_group.group import GroupModelForm
 from base.models.enums import education_group_categories
 from base.tests.factories.education_group_type import EducationGroupTypeFactory
 from base.tests.factories.education_group_year import EducationGroupYearFactory
