+++ base/views/education_groups/create.py
@@ -30,7 +30,7 @@ from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 from waffle.decorators import waffle_flag
 
-from base.forms.education_group.create import GroupForm
+from base.forms.education_group.group import GroupForm
 from base.models.education_group_year import EducationGroupYear
 from base.views import layout
 from base.views.common import display_success_messages, reverse_url_with_root
